requirements.txt
blinker==1.8.2
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
dominate==2.9.1
Flask==3.0.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
packaging==24.1
psycopg2-binary==2.9.10
pycparser==2.22
pytz==2024.2
visitor==0.1.3
Werkzeug==3.0.4
